more build stuff
