+++ window_background/background.js
@@ -22,61 +22,14 @@ const sha1 = require("js-sha1");
 
 const httpApi = require("./http-api");
 
-const db = require("../shared/database");
 const playerData = require("../shared/player-data");
 const { getReadableFormat } = require("../shared/util");
+const { HIDDEN_PW, MAIN_DECKS } = require("../shared/constants");
 const {
-  ARENA_MODE_MATCH,
-  ARENA_MODE_DRAFT,
-  DEFAULT_TILE,
-  HIDDEN_PW,
-  IPC_OVERLAY,
-  MAIN_DECKS
-} = require("../shared/constants");
-const {
-  getDateFormat,
   ipc_send,
   setData,
   unleakString,
-  parseWotcTimeFallback
 } = require("./background-util");
-const {
-  onLabelOutLogInfo,
-  onLabelGreToClient,
-  onLabelClientToMatchServiceMessageTypeClientToGREMessage,
-  onLabelInEventGetPlayerCourse,
-  onLabelInEventGetPlayerCourseV2,
-  onLabelInEventJoin,
-  onLabelInEventGetCombinedRankInfo,
-  onLabelInDeckGetDeckLists,
-  onLabelInDeckGetDeckListsV3,
-  onLabelInDeckGetPreconDecks,
-  onLabelInEventGetPlayerCourses,
-  onLabelInEventGetPlayerCoursesV2,
-  onLabelInDeckUpdateDeck,
-  onLabelInDeckUpdateDeckV3,
-  onLabelInventoryUpdated,
-  onLabelInPlayerInventoryGetPlayerInventory,
-  onLabelInPlayerInventoryGetPlayerCardsV3,
-  onLabelInProgressionGetPlayerProgress,
-  onLabelInEventDeckSubmit,
-  onLabelInEventDeckSubmitV3,
-  onLabelInEventGetActiveEvents,
-  onLabelEventMatchCreated,
-  onLabelOutDirectGameChallenge,
-  onLabelOutEventAIPractice,
-  onLabelInDraftDraftStatus,
-  onLabelInDraftMakePick,
-  onLabelOutDraftMakePick,
-  onLabelInEventCompleteDraft,
-  onLabelMatchGameRoomStateChangedEvent,
-  onLabelInEventGetSeasonAndRankDetail,
-  onLabelGetPlayerInventoryGetRewardSchedule,
-  onLabelRankUpdated,
-  onLabelMythicRatingUpdated,
-  onLabelTrackProgressUpdated,
-  onLabelTrackRewardTierUpdated
-} = require("./labels");
 
 const { createDeck } = require("./data");
 
@@ -93,11 +46,9 @@ var settingsStore = new Store({
 
 let logLoopInterval = null;
 const debugArenaID = undefined;
-var debugLogSpeed = 0.001;
 
 const addCustomDeck = require("./addCustomDeck");
 const forceDeckUpdate = require("./forceDeckUpdate");
-const getOpponentDeck = require("./getOpponentDeck");
 const {
   loadPlayerConfig,
   syncSettings,
@@ -161,7 +112,11 @@ ipc.on("start_background", function() {
   // app-level settings into singletons
 
   const appSettings = globals.rStore.get("settings");
-  const settings = { ...playerData.settings, ...appSettings, logUri };
+  const settings = {
+    ...playerData.settings,
+    ...appSettings,
+    logUri: globals.logUri
+  };
   setData({ settings }, false);
   ipc_send("initial_settings", settings);
 
@@ -435,311 +390,29 @@ ipc.on("set_log", function(event, arg) {
     globals.stopWatchingLog();
     globals.stopWatchingLog = startWatchingLog();
   }
-  logUri = arg;
+  globals.logUri = arg;
   settingsStore.set("logUri", arg);
 });
 
 // Read the log
 // Set variables to default first
-const mtgaLog = require("./mtga-log");
 let prevLogSize = 0;
 
-let logUri = mtgaLog.defaultLogUri();
 let settingsLogUri = settingsStore.get("logUri");
 if (settingsLogUri) {
-  logUri = settingsLogUri;
+  globals.logUri = settingsLogUri;
 }
 
 if (typeof process.env.LOGFILE !== "undefined") {
-  logUri = process.env.LOGFILE;
+  globals.logUri = process.env.LOGFILE;
 }
 
-console.log(logUri);
-const ArenaLogWatcher = require("./arena-log-watcher");
-
-let logReadStart = null;
-let logReadEnd = null;
+console.log(globals.logUri);
 
 function sendSettings() {
   let tags_colors = playerData.tags_colors;
   let settingsData = { tags_colors };
   httpApi.httpSetSettings(settingsData);
-}
-
-function onLogEntryFound(entry) {
-  if (globals.debugLog) {
-    let currentTime = new Date().getTime();
-    while (currentTime + debugLogSpeed >= new Date().getTime()) {
-      // sleep
-    }
-  }
-  let json;
-  if (entry.type == "connection") {
-    const data = {
-      arenaId: entry.socket.PlayerId,
-      arenaVersion: entry.socket.ClientVersion,
-      name: entry.socket.PlayerScreenName
-    };
-    setData(data);
-  } else if (entry.playerId && entry.playerId !== playerData.arenaId) {
-    return;
-  } else {
-    //console.log("Entry:", entry.label, entry, entry.json());
-    if (globals.firstPass) {
-      updateLoading(entry);
-    }
-    if (
-      (globals.firstPass && !playerData.settings.skip_firstpass) ||
-      !globals.firstPass
-    ) {
-      try {
-        switch (entry.label) {
-          case "Log.BI":
-          case "Log.Info":
-            if (entry.arrow == "==>") {
-              json = entry.json();
-              onLabelOutLogInfo(entry, json);
-            }
-            break;
-
-          case "GreToClientEvent":
-            json = entry.json();
-            onLabelGreToClient(entry, json);
-            break;
-
-          case "ClientToMatchServiceMessageType_ClientToGREMessage":
-            json = entry.json();
-            onLabelClientToMatchServiceMessageTypeClientToGREMessage(
-              entry,
-              json
-            );
-            break;
-
-          case "Event.GetPlayerCourse":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInEventGetPlayerCourse(entry, json);
-            }
-            break;
-
-          case "Event.GetPlayerCourseV2":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInEventGetPlayerCourseV2(entry, json);
-            }
-            break;
-
-          case "Event.Join":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInEventJoin(entry, json);
-            }
-            break;
-
-          case "Event.GetCombinedRankInfo":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInEventGetCombinedRankInfo(entry, json);
-            }
-            break;
-
-          case "Rank.Updated":
-            {
-              json = entry.json();
-              onLabelRankUpdated(entry, json);
-            }
-            break;
-
-          case "MythicRating.Updated":
-            {
-              json = entry.json();
-              onLabelMythicRatingUpdated(entry, json);
-            }
-            break;
-
-          case "Event.GetPlayerCourses":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInEventGetPlayerCourses(entry, json);
-            }
-            break;
-
-          case "Event.GetPlayerCoursesV2":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInEventGetPlayerCoursesV2(entry, json);
-            }
-            break;
-
-          case "Deck.GetDeckLists":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInDeckGetDeckLists(entry, json);
-            }
-            break;
-
-          case "Deck.GetDeckListsV3":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInDeckGetDeckListsV3(entry, json);
-            }
-            break;
-
-          case "Deck.GetPreconDecks":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInDeckGetPreconDecks(entry, json);
-            }
-            break;
-
-          case "Deck.UpdateDeck":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInDeckUpdateDeck(entry, json);
-            }
-            break;
-
-          case "Deck.UpdateDeckV3":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInDeckUpdateDeckV3(entry, json);
-            }
-            break;
-
-          case "Inventory.Updated":
-            json = entry.json();
-            onLabelInventoryUpdated(entry, json);
-            break;
-
-          case "PlayerInventory.GetPlayerInventory":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInPlayerInventoryGetPlayerInventory(entry, json);
-            }
-            break;
-
-          case "PlayerInventory.GetPlayerCardsV3":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInPlayerInventoryGetPlayerCardsV3(entry, json);
-            }
-            break;
-
-          case "Progression.GetPlayerProgress":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInProgressionGetPlayerProgress(entry, json);
-            }
-            break;
-
-          case "TrackProgress.Updated":
-            json = entry.json();
-            onLabelTrackProgressUpdated(entry, json);
-            break;
-
-          case "TrackRewardTier.Updated":
-            json = entry.json();
-            onLabelTrackRewardTierUpdated(entry, json);
-            break;
-
-          case "Event.DeckSubmit":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInEventDeckSubmit(entry, json);
-            }
-            break;
-
-          case "Event.DeckSubmitV3":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInEventDeckSubmitV3(entry, json);
-            }
-            break;
-
-          case "Event.MatchCreated":
-            json = entry.json();
-            onLabelEventMatchCreated(entry, json);
-            break;
-
-          case "Event.AIPractice":
-            if (entry.arrow == "==>") {
-              json = entry.json();
-              onLabelOutEventAIPractice(entry, json);
-            }
-            break;
-
-          case "DirectGame.Challenge":
-            if (entry.arrow == "==>") {
-              json = entry.json();
-              onLabelOutDirectGameChallenge(entry, json);
-            }
-            break;
-
-          case "Draft.DraftStatus":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInDraftDraftStatus(entry, json);
-            }
-            break;
-
-          case "Draft.MakePick":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInDraftMakePick(entry, json);
-            } else {
-              json = entry.json();
-              onLabelOutDraftMakePick(entry, json);
-            }
-            break;
-
-          case "Event.CompleteDraft":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInEventCompleteDraft(entry, json);
-            }
-            break;
-
-          case "Event.GetActiveEventsV2":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInEventGetActiveEvents(entry, json);
-            }
-            break;
-
-          case "MatchGameRoomStateChangedEvent":
-            json = entry.json();
-            onLabelMatchGameRoomStateChangedEvent(entry, json);
-            break;
-
-          case "Event.GetSeasonAndRankDetail":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelInEventGetSeasonAndRankDetail(entry, json);
-            }
-            break;
-
-          case "PlayerInventory.GetRewardSchedule":
-            if (entry.arrow == "<==") {
-              json = entry.json();
-              onLabelGetPlayerInventoryGetRewardSchedule(entry, json);
-            }
-            break;
-
-          default:
-            break;
-        }
-        if (entry.timestamp) {
-          setData({
-            last_log_timestamp: entry.timestamp,
-            last_log_format: getDateFormat(entry.timestamp)
-          });
-        }
-      } catch (err) {
-        console.log(entry.label, entry.position, entry.json());
-        console.error(err);
-      }
-    }
-  }
 }
 
 // Old parser
@@ -755,9 +428,9 @@ async function attemptLogLoop() {
 async function logLoop() {
   //console.log("logLoop() start");
   //ipc_send("ipc_log", "logLoop() start");
-  if (fs.existsSync(logUri)) {
-    if (fs.lstatSync(logUri).isDirectory()) {
-      ipc_send("no_log", logUri);
+  if (fs.existsSync(globals.logUri)) {
+    if (fs.lstatSync(globals.logUri).isDirectory()) {
+      ipc_send("no_log", globals.logUri);
       ipc_send("popup", {
         text: "No log file found. Please include the file name too.",
         time: 1000
@@ -765,7 +438,7 @@ async function logLoop() {
       return;
     }
   } else {
-    ipc_send("no_log", logUri);
+    ipc_send("no_log", globals.logUri);
     ipc_send("popup", { text: "No log file found.", time: 1000 });
     return;
   }
@@ -779,7 +452,7 @@ async function logLoop() {
   }
 */
 
-  const { size } = await mtgaLog.stat(logUri);
+  const { size } = await global.mtgaLog.stat(globals.logUri);
 
   if (size == undefined) {
     // Something went wrong obtaining the file size, try again later
@@ -795,8 +468,8 @@ async function logLoop() {
 
   const logSegment =
     delta > 0
-      ? await mtgaLog.readSegment(logUri, prevLogSize, delta)
-      : await mtgaLog.readSegment(logUri, 0, size);
+      ? await global.mtgaLog.readSegment(globals.logUri, prevLogSize, delta)
+      : await global.mtgaLog.readSegment(globals.logUri, 0, size);
 
   // We are looping only to get user data (processLogUser)
   // Process only the user data for initial loading (prior to log in)
@@ -925,65 +598,4 @@ function dataChop(data, startStr, endStr) {
   }
 
   return data;
-}
-
-//
-function updateLoading(entry) {
-  if (globals.firstPass) {
-    const completion = entry.position / entry.size;
-    ipc_send("popup", {
-      text: `Reading log: ${Math.round(100 * completion)}%`,
-      time: 0,
-      progress: completion
-    });
-  }
-}
-
-//
-function finishLoading() {
-  if (globals.firstPass) {
-    ipc_send("popup", {
-      text: "Finishing initial log read...",
-      time: 0,
-      progress: 2
-    });
-    globals.firstPass = false;
-    globals.store.set(playerData.data);
-    logReadEnd = new Date();
-    let logReadElapsed = (logReadEnd - logReadStart) / 1000;
-    ipc_send("ipc_log", `Log read in ${logReadElapsed}s`);
-
-    ipc_send("popup", {
-      text: "Initializing...",
-      time: 0,
-      progress: 2
-    });
-
-    if (globals.duringMatch) {
-      ipc_send("set_arena_state", ARENA_MODE_MATCH);
-      update_deck(false);
-    } else if (globals.duringDraft) {
-      ipc_send("set_arena_state", ARENA_MODE_DRAFT);
-    }
-
-    ipc_send("set_settings", JSON.stringify(playerData.settings));
-    ipc_send("initialize");
-    ipc_send("player_data_refresh");
-
-    if (playerData.name) {
-      httpApi.httpSetPlayer(
-        playerData.name,
-        playerData.rank.constructed.rank,
-        playerData.rank.constructed.tier,
-        playerData.rank.limited.rank,
-        playerData.rank.limited.tier
-      );
-    }
-
-    ipc_send("popup", {
-      text: "Initialized successfully!",
-      time: 3000,
-      progress: -1
-    });
-  }
 }
